+++ tests/testthat/test_plots.R
@@ -102,4 +102,5 @@ test_that("plots look the way they should", {
          values = c(1.0,0.8,0.6,0.4,0.2,0)
        ) 
   })
+
 })